+++ packages/dialog/src/dialog.vue
@@ -1,24 +1,31 @@
 <template>
   <div :style="visible ? 'opacity: 1;' : 'opacity: 0; display: none;'">
     <div class="weui-mask"></div>
-    <div :class="{'weui-skin_android': type == 'android'}" class="weui-dialog">
+    <div
+      :class="{ 'weui-skin_android': type == 'android' }"
+      class="weui-dialog"
+    >
       <div class="weui-dialog__hd" v-if="title">
-        <strong class="weui-dialog__title">{{title}}</strong>
+        <strong class="weui-dialog__title">{{ title }}</strong>
       </div>
-      <div class="weui-dialog__bd" v-html="content"></div>
+      <div class="weui-dialog__bd">
+        <slot></slot>
+      </div>
       <div class="weui-dialog__ft">
         <a
           @click="cancel()"
           class="weui-dialog__btn weui-dialog__btn_default"
           href="javascript:"
           v-if="cancelText || showCancel"
-        >{{cancelText}}</a>
+          >{{ cancelText }}</a
+        >
         <a
           @click="ok()"
           class="weui-dialog__btn weui-dialog__btn_primary"
           href="javascript:"
           v-if="okText || showOk"
-        >{{okText}}</a>
+          >{{ okText }}</a
+        >
       </div>
     </div>
   </div>
@@ -46,10 +53,6 @@ export default {
     title: {
       type: String,
       default: '弹窗标题'
-    },
-    content: {
-      type: String,
-      default: '弹窗内容'
     },
     cancelText: {
       type: String,